+++ Assets/Scripts/Hud/HealthBar.cs
@@ -1,18 +1,17 @@
 ﻿using UnityEngine;
 using UnityEngine.UI;
-using TMPro;
 
 public class HealthBar : MonoBehaviour
 {
 
     public Slider slider;
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
 
-    public void SetMaxHealth(int value)
+    public void SetMaxHealth(float value)
     {
         slider.maxValue = value;
     }
